backend/requirements.txt
boto3==1.40.37
botocore==1.40.37
Django==5.2.6
django-cors-headers==4.4.0
djangorestframework==3.16.1
google-generativeai==0.7.1
PyMuPDF==1.26.4
python-docx==1.2.0
qdrant-client==1.15.1
sentence-transformers==5.1.1